UPSTREAM: tun: implement header_ops->parse_protocol for AF_PACKET
The tun driver passes up skb->protocol to userspace in the form of PI headers.
For AF_PACKET injection, we need to support its call chain of:

    packet_sendmsg -> packet_snd -> packet_parse_headers ->
      dev_parse_header_protocol -> parse_protocol

Without a valid parse_protocol, this returns zero, and the tun driver
then gives userspace bogus values that it can't deal with.

Note that this isn't the case with tap, because tap already benefits
from the shared infrastructure for ethernet headers. But with tun,
there's nothing.

Change-Id: I1505a341bf1d5cd2ae9ad8a78492c8c8a55ee817
Signed-off-by: Kunmun <kunmun.devroms@gmail.com>
Signed-off-by: RyuujiX <saputradenny712@gmail.com>
Signed-off-by: nerd.zenss_ <nerd.zensdev@gmail.com>